Flask_App/requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
appscript==1.2.5
astunparse==1.6.3
attrs==23.2.0
audioread==3.0.1
Babel==2.15.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
docx2pdf==0.1.8
docxcompose==1.4.0
docxtpl==0.17.0
filelock==3.15.3
Flask==3.0.3
Flask-Mail==0.10.0
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.0
gast==0.5.4
google-auth==2.30.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.64.1
h5py==3.11.0
huggingface-hub==0.23.4
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
keras==2.15.0
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2
llvmlite==0.43.0
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
namex==0.0.8
networkx==3.3
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
opencv-contrib-python==4.10.0.84
opt-einsum==3.3.0
optree==0.11.0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.2
pooch==1.8.2
protobuf==4.25.3
psutil==6.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
SpeechRecognition==3.10.4
sympy==1.12.1
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io==0.37.0
tensorflow-io-gcs-filesystem==0.37.0
tensorflow-macos==2.15.0
termcolor==2.4.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.1
torch_geometric==2.5.3
tqdm==4.66.4
transformers==4.41.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
wfdb==4.1.2
wrapt==1.14.1
yarl==1.9.4
